refactor(domain): separate fare and path domains
- delete unused files
