+++ src/main/java/wooteco/subway/domain/fare/AgeDiscountFare.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.path.fare;
+package wooteco.subway.domain.fare;
 
 public class AgeDiscountFare extends Decorator {
