+++ src/main/java/wooteco/subway/domain/fare/DistanceOverFare.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.path.fare;
+package wooteco.subway.domain.fare;
 
 public class DistanceOverFare extends Decorator {
