+++ src/main/java/wooteco/subway/domain/fare/LineOverFare.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.path.fare;
+package wooteco.subway.domain.fare;
 
 import java.util.Comparator;
 import java.util.List;
