+++ src/main/java/wooteco/subway/domain/path/NavigatorJgraphtAdapter.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.path2;
+package wooteco.subway.domain.path;
 
 import java.util.ArrayList;
 import java.util.HashSet;
