+++ src/main/java/wooteco/subway/service/PathService.java
@@ -3,12 +3,12 @@ package wooteco.subway.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import wooteco.subway.domain.line.LineInfo;
-import wooteco.subway.domain.path2.Path;
-import wooteco.subway.domain.path.fare.AgeDiscountFare;
-import wooteco.subway.domain.path.fare.BasicFare;
-import wooteco.subway.domain.path.fare.DistanceOverFare;
-import wooteco.subway.domain.path.fare.Fare;
-import wooteco.subway.domain.path.fare.LineOverFare;
+import wooteco.subway.domain.path.Path;
+import wooteco.subway.domain.fare.AgeDiscountFare;
+import wooteco.subway.domain.fare.BasicFare;
+import wooteco.subway.domain.fare.DistanceOverFare;
+import wooteco.subway.domain.fare.Fare;
+import wooteco.subway.domain.fare.LineOverFare;
 import wooteco.subway.domain.section.Section;
 import wooteco.subway.domain.station.Station;
 import wooteco.subway.dto.response.PathResponse;
